+++ src/pharmpy/modeling/odes.py
@@ -130,6 +130,8 @@ def set_zero_order_elimination(model):
     set_michaelis_menten_elimination
 
     """
+    if has_zero_order_elimination(model):
+        return model
     _do_michaelis_menten_elimination(model)
     obs = get_observations(model)
     init = obs.min() / 100  # 1% of smallest observation
@@ -151,8 +153,8 @@ def has_michaelis_menten_elimination(model):
 
     Return
     ------
-    Model
-        Reference to the same model
+    bool
+        True if model has describes Michaelis-Menten elimination
 
     Examples
     --------
@@ -174,6 +176,44 @@ def has_michaelis_menten_elimination(model):
     is_zero_order = 'POP_KM' in model.parameters and model.parameters['POP_KM'].fix
     could_be_mixed = sympy.Symbol('CL') in rate.free_symbols
     return is_nonlinear and not is_zero_order and not could_be_mixed
+
+
+def has_zero_order_elimination(model):
+    """Check if the model describes zero-order elimination
+
+    This function relies on heuristics and will not be able to detect all
+    possible ways of coding the zero-order elimination.
+
+    Parameters
+    ----------
+    model : Model
+        Pharmpy model
+
+    Return
+    ------
+    bool
+        True if model has describes zero order elimination
+
+    Examples
+    --------
+    >>> from pharmpy.modeling import *
+    >>> model = load_example_model("pheno")
+    >>> has_zero_order_elimination(model)
+    False
+    >>> set_zero_order_elimination(model)   # doctest: +ELLIPSIS
+    <...>
+    >>> has_zero_order_elimination(model)
+    True
+    """
+    model.statements.to_compartmental_system()
+    odes = model.statements.ode_system
+    output = odes.output_compartment
+    central = odes.central_compartment
+    rate = odes.get_flow(central, output)
+    is_nonlinear = odes.t in rate.free_symbols
+    is_zero_order = 'POP_KM' in model.parameters and model.parameters['POP_KM'].fix
+    could_be_mixed = sympy.Symbol('CL') in rate.free_symbols
+    return is_nonlinear and is_zero_order and not could_be_mixed
 
 
 def set_michaelis_menten_elimination(model):